✨ feat:
任务栏图标+快捷键功能提升至index页面、快捷键区分press和release优化使用、任务栏登录与未登录显示不同的menu、应用开机自启功能（注册表+改变当前工作目录）

+++ tools/fileOperate.py
@@ -106,12 +106,13 @@ class File(object):
         exe_dir = os.path.dirname(exe_file)
         return exe_dir
 
-    def delete_file(self, file_path):
-        try:
+    def delete_file(self, path_list):
+        # 构建文件路径
+        file_path = os.path.join(*path_list)
+        # 如果文件路径存在，则删除
+        if os.path.exists(file_path):
             os.remove(file_path)
-            debugLog(f"文件 {file_path} 删除成功")
-        except Exception as e:
-            debugLog(f"删除文件 {file_path} 失败：{e}")
+            debugLog(f"{file_path}已成功移除")
 
     def delete_files_with_name(self, directory, name):
         # 遍历目录中的文件
